+++ dokan-class.php
@@ -25,7 +25,7 @@ final class WeDevs_Dokan {
      *
      * @var string
      */
-    public $version = '4.2.10';
+    public $version = '4.3.0';
 
     /**
      * Instance of self